Move ParameterScaler to instrument/parameter.py
